iter as pinned-vec method
`iter` is implemented as a `SplitVec` method rather than `FixedVec` method. Documentation is updated accordingly.
