+++ src/common_traits/mod.rs
@@ -3,4 +3,3 @@ mod debug;
 mod eq;
 mod index;
 mod into_iter;
-mod iter;
